Refactor HybridSearchRatio from float64 to int percentage
Change HybridSearchRatio from *float64 (0.0-1.0) to *int (0-100 percentage)
to avoid needing allowDangerousTypes=true in controller-gen.

- Update type definitions in config, optimizer, and server packages
- Update conversion logic in hybrid.go to convert percentage to ratio
- Update all test files with new percentage values
- Update config files, examples, and documentation
- Remove allowDangerousTypes=true from Taskfile.yml

This is a breaking change: users need to update configs from 0.7 to 70, etc.

+++ pkg/vmcp/optimizer/optimizer.go
@@ -51,8 +51,8 @@ type Config struct {
 	// (empty = auto-default: ":memory:" or "{PersistPath}/fts.db")
 	FTSDBPath string
 
-	// HybridSearchRatio controls semantic vs BM25 mix (0.0-1.0, default: 0.7)
-	HybridSearchRatio float64
+	// HybridSearchRatio controls semantic vs BM25 mix (0-100 percentage, default: 70)
+	HybridSearchRatio int
 
 	// EmbeddingConfig configures the embedding backend (vLLM, Ollama, placeholder)
 	EmbeddingConfig *embeddings.Config
